+++ examples/User/Action/GetUsersAction.php
@@ -15,7 +15,7 @@ class GetUsersAction
     public function __invoke(JsonApi $jsonApi): ResponseInterface
     {
         // Extracting pagination information from the request, page = 1, size = 10 if it is missing
-        $pagination = $jsonApi->getRequest()->getPageBasedPagination(1, 10);
+        $pagination = $jsonApi->getPaginationFactory()->createPageBasedPagination(1, 10);
 
         // Retrieving a paginated collection of user domain objects
         $users = UserRepository::getUsers($pagination->getPage(), $pagination->getSize());