+++ tests/JsonApi/Request/Pagination/PaginationFactoryTest.php
@@ -1,0 +1,112 @@
+<?php
+declare(strict_types=1);
+
+namespace WoohooLabs\Yin\Tests\JsonApi\Request\Pagination;
+
+use PHPUnit\Framework\TestCase;
+use WoohooLabs\Yin\JsonApi\Exception\DefaultExceptionFactory;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\CursorBasedPagination;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\FixedCursorBasedPagination;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\FixedPageBasedPagination;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\OffsetBasedPagination;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\PageBasedPagination;
+use WoohooLabs\Yin\JsonApi\Request\Pagination\PaginationFactory;
+use WoohooLabs\Yin\JsonApi\Request\Request;
+use WoohooLabs\Yin\JsonApi\Serializer\JsonDeserializer;
+use Zend\Diactoros\ServerRequest;
+
+class PaginationFactoryTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function createFixedPageBasedPagination()
+    {
+        $paginationFactory = $this->createPaginationFactoryFromRequestQueryParams(
+            [
+                "page" => ["number" => 1],
+            ]
+        );
+
+        $pagination = $paginationFactory->createFixedPageBasedPagination();
+
+        $this->assertEquals(new FixedPageBasedPagination(1), $pagination);
+    }
+
+    /**
+     * @test
+     */
+    public function createPageBasedPagination()
+    {
+        $paginationFactory = $this->createPaginationFactoryFromRequestQueryParams(
+            [
+                "page" => ["number" => 1, "size" => 10],
+            ]
+        );
+
+        $pagination = $paginationFactory->createPageBasedPagination();
+
+        $this->assertEquals(new PageBasedPagination(1, 10), $pagination);
+    }
+
+    /**
+     * @test
+     */
+    public function createOffsetBasedPagination()
+    {
+        $paginationFactory = $this->createPaginationFactoryFromRequestQueryParams(
+            [
+                "page" => ["offset" => 1, "limit" => 10],
+            ]
+        );
+
+        $pagination = $paginationFactory->createOffsetBasedPagination();
+
+        $this->assertEquals(new OffsetBasedPagination(1, 10), $pagination);
+    }
+
+    /**
+     * @test
+     */
+    public function createCursorBasedPagination()
+    {
+        $paginationFactory = $this->createPaginationFactoryFromRequestQueryParams(
+            [
+                "page" => ["cursor" => "abc", "size" => 10],
+            ]
+        );
+
+        $pagination = $paginationFactory->createCursorBasedPagination();
+
+        $this->assertEquals(new CursorBasedPagination("abc", 10), $pagination);
+    }
+
+    /**
+     * @test
+     */
+    public function createFixedCursorBasedPagination()
+    {
+        $paginationFactory = $this->createPaginationFactoryFromRequestQueryParams(
+            [
+                "page" => ["cursor" => "abc"],
+            ]
+        );
+
+        $pagination = $paginationFactory->createFixedCursorBasedPagination();
+
+        $this->assertEquals(new FixedCursorBasedPagination("abc"), $pagination);
+    }
+
+    private function createPaginationFactoryFromRequestQueryParams(array $queryParams): PaginationFactory
+    {
+        return new PaginationFactory($this->createRequestWithQueryParams($queryParams));
+    }
+
+    private function createRequestWithQueryParams(array $queryParams): Request
+    {
+        $psrRequest = new ServerRequest();
+        $psrRequest = $psrRequest->withQueryParams($queryParams);
+
+        return new Request($psrRequest, new DefaultExceptionFactory(), new JsonDeserializer());
+    }
+}